[refactor] :truck: add 'Model' suffix to models for consistency across layers (#276)
