+++ src/models/nftModel.ts
